Modularize JS code (#2822)
* Migrate active_units/modal

* modularize openpose editor

* modularize photopea editor
